fix route not refreshing on sign in
