+++ apps/client/app/(navbar)/ClientButtons.tsx
@@ -1,6 +1,8 @@
 "use client";
 
+import { useRouter } from "next/navigation";
 import { SessionProvider } from "next-auth/react";
+import { useEffect, useTransition } from "react";
 
 import { useSession } from "../../src/client/auth/useSession";
 import ProfileButton from "./ProfileButton";
@@ -15,6 +17,23 @@ export default function ClientButtons() {
 }
 
 function Buttons() {
-  const { data: session } = useSession();
-  return session?.address ? <ProfileButton /> : <SignInButton />;
+  const { data: session, status } = useSession();
+  const router = useRouter();
+  const [transitionLoading, startTransition] = useTransition();
+
+  useEffect(() => {
+    if (status === "loading") return;
+
+    startTransition(() => {
+      router.refresh();
+    });
+  }, [status, router, startTransition]);
+
+  const isLoading = transitionLoading || status === "loading";
+
+  return session?.address ? (
+    <ProfileButton isLoading={isLoading} />
+  ) : (
+    <SignInButton isLoading={isLoading} />
+  );
 }
